pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dspy-coding-agent"
version = "0.1.0"
description = "A DSPy-based autonomous coding agent with comprehensive tooling"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dspy>=2.5",
    "diff-match-patch>=20230430",
    "rope>=1.13",
    "swe-rex>=0.2",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
mcp = [
    "mcp>=1.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

